refactor(www): give each create axis its own folder
Make the customizer module uniform: every design-system axis is now a
folder whose `index.tsx` is the entry, instead of a mix of flat
`*-config.tsx` files and a couple of folders.

  create/
    colors/      index.tsx + knobs.tsx + contrast.tsx + contrast.spec.ts
    typography/  index.tsx
    iconography/ index.tsx
    layout/      index.tsx   (radius + density)
    cursor/      index.tsx
    components/  index.tsx
    code-options/ index.tsx + controls.tsx + preview.tsx

Pure move — git tracks every file as a rename, no behaviour change.
code-options drops its barrel (dialog.tsx → index.tsx). The export
actions (install-command, open-in-v0), preset/ and preview/ are left as
they are. customizer-panel.tsx imports update to the folder paths.
