--- www/src/modules/create/code-options/index.ts
@@ -1,1 +1,0 @@
-export { CodeOptionsDialog } from './dialog'
